Refactor checker and test checker added
Expected errors
ERROR: Illegal use of literal for property: background-color, it must be a ColorLiteral.
ERROR: Illegal use of variable: Width, it is undefined or cant be used in current scope
ERROR: Illegal use of literal for property: width, a size property should use Percentage or PixelLiteral.
ERROR: Illegal use of literal for property: color, it must be a ColorLiteral.
ERROR: Illegal use of ColorLiteral in math operation
ERROR: Illegal property used: something, Only (color, background-color, width, height) are allowed.
ERROR: Illegal use of literal for property: color, it must be a ColorLiteral.
ERROR: Illegal use of variable: NewColor, it is undefined or cant be used in current scope
ERROR: Illegal IfClause condition: Percentage is not a Boolean